Change Employee ID type from Long to Integer

+++ EmployeeRepository.java
@@ -3,11 +3,9 @@ package com.example.demo.repository;
 import com.example.demo.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-    // Optional: list employees by designation
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
     List<Employee> findByDesignation(String designation);
 }
